admin: refactor parameter values list to avoid files N+1 problem
- list now displays one "preview" column with file thumbnail if exists, or the rgb hex color background if file does not exist
- FileApiRepository::getAllFilesIndexedByEntityId() moved to UploadedFileRepository in the framework package
