+++ packages/administration/templates/content/parameterValue/listGrid.html.twig
@@ -4,27 +4,22 @@
     <a href="{{ url('admin_parametervalue_edit', { id: row.pv.id }) }}">{{ value }}</a>
 {% endblock %}
 
-{% block grid_value_cell_id_rgbHex %}
-    {% if row.pv.rgbHex %}
-        {{ row.pv.rgbHex }}
-    {% else %}
-        {% set uploadedFiles = uploadedFilesByEntityNameAndId('parameterValue', row.pv.id) %}
-        {% if uploadedFiles|length > 0 %}
-            <span class="text-muted">{{ 'Custom image'|trans }}</span>
-        {% else %}
-            <span>-</span>
-        {% endif %}
-    {% endif %}
-{% endblock %}
-
-{% block grid_value_cell_id_colourIcon %}
-    {% set uploadedFiles = uploadedFilesByEntityNameAndId('parameterValue', row.pv.id) %}
-    {% if uploadedFiles|length > 0 %}
-        <img src="{{ uploadedFileUrl(uploadedFiles[0]) }}"
+{% block grid_value_cell_id_preview %}
+    {% set uploadedFiles = filesIndexedByParameterValueIds[row.pv.id]|default([]) %}
+    {% if uploadedFiles is not empty %}
+        {% set uploadedFile = uploadedFiles|first %}
+        <img src="{{ uploadedFileUrl(uploadedFile) }}"
+             title="{{ uploadedFile.name }}"
+             data-bs-toggle="tooltip"
+             data-bs-placement="right"
              alt="{{ row.pv.text }}"
              style="width: 30px; height: 30px; object-fit: cover; border-radius: 4px;">
     {% elseif row.pv.rgbHex %}
-        <div style="width: 30px; height: 30px; background-color: {{ row.pv.rgbHex }}; border-radius: 4px; border: 1px solid #ccc;"></div>
+        <div style="width: 30px; height: 30px; background-color: {{ row.pv.rgbHex }}; border-radius: 4px; border: 1px solid #ccc;"
+             title="{{ row.pv.rgbHex }}"
+             data-bs-toggle="tooltip"
+             data-bs-placement="right"
+        ></div>
     {% else %}
         <span>-</span>
     {% endif %}
