+++ packages/frontend-api/src/Model/Resolver/Products/ParameterValueFilesQuery.php
@@ -7,14 +7,13 @@ namespace Shopsys\FrontendApiBundle\Model\Resolver\Products;
 use GraphQL\Executor\Promise\Promise;
 use Overblog\DataLoader\DataLoaderInterface;
 use Shopsys\FrameworkBundle\Component\UploadedFile\Config\UploadedFileTypeConfig;
+use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue;
 use Shopsys\FrontendApiBundle\Component\Files\FileBatchLoadData;
 use Shopsys\FrontendApiBundle\Model\Product\Filter\ParameterValueFilterOption;
 use Shopsys\FrontendApiBundle\Model\Resolver\AbstractQuery;
 
 class ParameterValueFilesQuery extends AbstractQuery
 {
-    public const string PARAMETER_VALUE_ENTITY_NAME = 'parameterValue';
-
     /**
      * @param \Overblog\DataLoader\DataLoaderInterface $filesBatchLoader
      */
@@ -33,7 +32,7 @@ class ParameterValueFilesQuery extends AbstractQuery
         return $this->filesBatchLoader->load(
             new FileBatchLoadData(
                 $parameterValueFilterOption->parameterValue->getId(),
-                static::PARAMETER_VALUE_ENTITY_NAME,
+                ParameterValue::ENTITY_NAME_FOR_FILES_CONFIG,
                 UploadedFileTypeConfig::DEFAULT_TYPE_NAME,
             ),
         );
